requirements-new.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# TUI Framework
textual>=0.47.0
rich>=13.0.0

# Utilities
termcolor>=2.3.0

# AI Integration
openai>=1.0.0
requests>=2.31.0

# Optional dependencies for enhanced features
# Install with: pip install -r requirements-new.txt

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# mypy>=1.5.0
